pyproject.toml
[project]
name = "scisciDB"
requires-python = ">=3.12"
version = "0.1.0"
dependencies = [
    "dagster==1.12.21",
    "dagster-slurm",
    "duckdb>=1.5.1",
    "python-dotenv",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }

[tool.uv.sources]
dagster-slurm = { path = "../dagster-slurm/projects/dagster-slurm", editable = true }

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "scisciDB"
registry_modules = [
    "scisciDB.components.*",
]


